% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius to Fahrenheit}
\usage{
C_to_F(C_temp)
}
\arguments{
\item{C_temp}{A numeric vector of temperatures in Celsius.}
}
\value{
A numeric vector of temperatures in Fahrenheit.
}
\description{
Convert temperatures from Celsius to Fahrenheit.
}
\examples{
C_to_F(0)
C_to_F(c(10, 20, 30))
}
