% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{C_to_K}
\alias{C_to_K}
\title{Celsius to Kelvin}
\usage{
C_to_K(C_temp)
}
\arguments{
\item{C_temp}{A numeric vector of temperatures in Celsius.}
}
\value{
A numeric vector of temperatures in Kelvin.
}
\description{
Convert temperatures from Celsius to Kelvin.
}
\examples{
C_to_K(0)
C_to_K(c(10, 20, 30))
}
