% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit to Celsius}
\usage{
F_to_C(F_temp)
}
\arguments{
\item{F_temp}{A numeric vector of temperatures in Fahrenheit.}
}
\value{
A numeric vector of temperatures in Celsius.
}
\description{
Convert temperatures from Fahrenheit to Celsius.
}
\examples{
F_to_C(32)
F_to_C(c(50, 68, 104))
}
