% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{F_to_K}
\alias{F_to_K}
\title{Fahrenheit to Kelvin}
\usage{
F_to_K(F_temp)
}
\arguments{
\item{F_temp}{A numeric vector of temperatures in Fahrenheit.}
}
\value{
A numeric vector of temperatures in Kelvin.
}
\description{
Convert temperatures from Fahrenheit to Kelvin.
}
\examples{
F_to_K(32)
F_to_K(c(50, 68, 104))
}
