% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_C}
\alias{K_to_C}
\title{Kelvin to Celsius}
\usage{
K_to_C(K_temp)
}
\arguments{
\item{K_temp}{A numeric vector of temperatures in Kelvin.}
}
\value{
A numeric vector of temperatures in Celsius.
}
\description{
Convert temperatures from Kelvin to Celsius.
}
\examples{
K_to_C(273.15)
K_to_C(c(283.15, 293.15, 303.15))
}
