% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_conversion.R
\name{K_to_F}
\alias{K_to_F}
\title{Kelvin to Fahrenheit}
\usage{
K_to_F(K_temp)
}
\arguments{
\item{K_temp}{A numeric vector of temperatures in Kelvin.}
}
\value{
A numeric vector of temperatures in Fahrenheit.
}
\description{
Convert temperatures from Kelvin to Fahrenheit.
}
\examples{
K_to_F(273.15)
K_to_F(c(283.15, 293.15, 303.15))
}
